+++ Sources/ProcessOut/Sources/Api/ProcessOutApi.swift
@@ -91,7 +91,7 @@ private final class SharedProcessOutApi: ProcessOutApiType {
     private lazy var repositoryLogger = createLogger(for: Constants.repositoryLoggerCategory)
 
     private lazy var httpConnector: HttpConnectorType = {
-        let connectorConfiguration = HttpConnector.Configuration(
+        let connectorConfiguration = HttpConnectorConfiguration(
             baseUrl: configuration.apiBaseUrl,
             version: Self.version
         )
