+++ Sources/ProcessOut/Sources/Connectors/Http/Models/HttpConnectorFailure.swift
